Add `thumbFocusCapture` prop to `use-slider`
Adds a boolean `thumbFocusCapture` to `use-slider` so users can opt-out of the default behavior where the slider thumb captures focus on every value change - expected use-case is with a controlled slider that shares a page with other elements require focus for user input.  Without it the slider will continually "steal" focus by default.